requirements.txt
requests>=2.28
beautifulsoup4>=4.12
feedparser>=6.0
aiohttp>=3.9
lxml>=4.9
urllib3>=2.0
pydantic>=2.6
trafilatura>=1.7
PyYAML>=6.0
pytest>=7.4
fastapi>=0.110
uvicorn[standard]>=0.27
asyncpg>=0.29
PyMuPDF>=1.23
tenacity>=8.2
